Rename BehaviorNode::Cooldown to BehaviorNode::Deadline and introduce a new BehaviorNode::Cooldown
Deadline expires at a fixed Instant, calculated when the node is
(re-)initialized. Cooldown expires after a given duration, and that
duration is only decremented when the node is evaluated.

As an example, if a cooldown is wrapped in a Condition node, it will
only decrease when the Condition is true (the initial use-case is
creature respawn delay which should only decrease when the creature is
dead).